requirements.txt
pandas
numpy
matplotlib
scikit-learn
emoji
nltk
openpyxl
seaborn
PySastrawi
wordcloud
plotly